requirements.txt
black==22.6.0
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
cycler==0.11.0
docker-pycreds==0.4.0
fonttools==4.34.4
gitdb==4.0.9
GitPython==3.1.27
idna==3.3
imageio==2.21.1
imageio-ffmpeg==0.4.7
joblib==1.1.0
kiwisolver==1.4.4
matplotlib==3.5.2
mypy-extensions==0.4.3
numpy==1.23.1
opencv-python==4.6.0.66
packaging==21.3
pandas==1.4.3
pathspec==0.9.0
pathtools==0.1.2
Pillow==9.2.0
platformdirs==2.5.2
promise==2.3
protobuf==3.20.1
psutil==5.9.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
requests==2.28.1
scikit-learn==1.1.2
scipy==1.9.0
sentry-sdk==1.9.3
setproctitle==1.3.1
shortuuid==1.0.9
six==1.16.0
smmap==5.0.0
threadpoolctl==3.1.0
timm==0.6.7
tomli==2.0.1
torch==1.12.1
torchaudio==0.12.1
torchvision==0.13.1
--extra-index-url https://download.pytorch.org/whl/cu116 
tqdm==4.64.0
typing-extensions==4.3.0
urllib3==1.26.11
wandb==0.13.1
